feat: add AuctionPublicItem as separate component.
