+++ dapp/src/app/(protected)/auction/page.tsx
@@ -3,48 +3,24 @@
 
 'use client';
 
-import { Clock, FilterList, IotaLogoSmall, Search } from '@iota/apps-ui-icons';
+import { FilterList, Search } from '@iota/apps-ui-icons';
 import {
-    // Badge,
-    // BadgeType,
     Button,
     ButtonSegment,
     ButtonSegmentType,
     ButtonType,
-    Divider,
-    DividerType,
     Dropdown,
     Input,
     InputType,
     ListItem,
-    // InfoBox,
-    // InfoBoxStyle,
-    // InfoBoxType,
-    // LoadingIndicator,
     SegmentedButton,
 } from '@iota/apps-ui-kit';
-import { normalizeIotaName } from '@iota/iota-names-sdk';
-// import { formatBalance } from '@iota/iota-sdk/utils';
 import { useEffect, useMemo, useRef, useState } from 'react';
 
-import {
-    AuctionBidDialog,
-    AuctionDetails,
-    formatTimeRemaining,
-    getTimeRemaining,
-    // useGetUserAuctions,
-    // useGetUserAuctions
-} from '@/auctions';
+import { AuctionBidDialog, AuctionDetails } from '@/auctions';
+import { AuctionPublicItem } from '@/auctions/components/AuctionPublicItem';
 import { useAuctions } from '@/auctions/hooks/useAuctions';
-import { useCountdown } from '@/auctions/hooks/useCountdown';
-import { NameCard } from '@/components/name-card/NameCard';
-import { NameCardBody } from '@/components/name-card/NameCardBody';
 import { useDebounce } from '@/hooks/useDebounce';
-import { formatNanosToIota } from '@/lib/utils';
-
-// import { AuctionItem } from '@/auctions/components/AuctionItem';
-// import { SearchInput } from '@/components/layout/SearchInput';
-// import { AvatarDisplay } from '@/components/name-record/AvatarDisplay';
 
 enum TypeFilter {
     Active = 'Active',
@@ -142,10 +118,6 @@ export default function AuctionsPage(): JSX.Element {
     const displayedAuctions =
         selectedFilter === TypeFilter.Active ? activeAuctions : finishedAuctions;
 
-    console.log('auctions', auctions);
-    console.log('active auctions', activeAuctions);
-    console.log('finished auctions', finishedAuctions);
-
     return (
         <>
             <div className="flex flex-row gap-md items-center pt-[80px] md:pt-0">
@@ -210,19 +182,7 @@ export default function AuctionsPage(): JSX.Element {
             <div className="mt-8 flex flex-wrap">
                 {displayedAuctions.map((auction) => (
                     <div key={auction.name} className="w-1/3 md:w-1/4 p-2">
-                        <NameCard name={auction.name} size="full">
-                            <NameCardBody name={normalizeIotaName(auction.name)}>
-                                <AuctionBidPrice
-                                    price={auction.metadata?.currentBidNanos}
-                                    auctionName={auction.name}
-                                    onBidClick={setBidDialogName}
-                                />
-                                <div className="my-4">
-                                    <Divider type={DividerType.Horizontal} />
-                                </div>
-                                <AuctionTimeRemaining auction={auction} />
-                            </NameCardBody>
-                        </NameCard>
+                        <AuctionPublicItem auction={auction} onBidClick={setBidDialogName} />
                     </div>
                 ))}
             </div>
@@ -234,70 +194,9 @@ export default function AuctionsPage(): JSX.Element {
                     closeDialog={() => setBidDialogName(null)}
                     onCompleted={() => {
                         setBidDialogName(null);
-                        // Optionally refresh auctions data
                     }}
                 />
             )}
         </>
-    );
-}
-
-function AuctionBidPrice({
-    price,
-    auctionName,
-    onBidClick,
-}: {
-    price?: bigint;
-    auctionName: string;
-    onBidClick: (name: string) => void;
-}): JSX.Element {
-    const formattedPrice = formatNanosToIota(price ?? BigInt(0));
-
-    return (
-        <div className="flex items-center justify-between">
-            <div>
-                <div className="text-label-md text-names-neutral-50">Current Bid</div>
-                <div className="flex mt-1 items-center gap-2">
-                    <div className="bg-names-solid-blue rounded-full w-5 h-5 flex items-center justify-center">
-                        <IotaLogoSmall className="w-4 h-4" />
-                    </div>
-                    <div className="flex items-baseline gap-1">
-                        <span className="text-body-lg">{formattedPrice}</span>
-                        {/* <span className="text-label-md text-names-neutral-70">
-                            ${convertedToFiat.toString()}
-                        </span> */}
-                    </div>
-                </div>
-            </div>
-            <div>
-                <Button
-                    text="Bid"
-                    onClick={() => onBidClick(auctionName)}
-                    fullWidth
-                    type={ButtonType.Outlined}
-                />
-            </div>
-        </div>
-    );
-}
-
-function AuctionTimeRemaining({ auction }: { auction: AuctionDetails }) {
-    // Always call hooks at the top level - never inside conditions
-    const timeRemainingMs = getTimeRemaining(auction.metadata);
-    const { milliseconds } = useCountdown(timeRemainingMs);
-    const formattedTimeRemaining = formatTimeRemaining(milliseconds);
-
-    return (
-        <div className="flex items-center justify-between">
-            <div>
-                <span className="text-label-md text-names-neutral-50">Time left</span>
-            </div>
-            <div className="flex items-center gap-1">
-                <Clock className="w-4 h-4 text-names-neutral-50" />
-                <span className="text-label-md text-names-neutral-50">
-                    {formattedTimeRemaining}
-                </span>
-            </div>
-        </div>
     );
 }
